pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "reproducing-ttt-e2e"
version = "0.1.0"
description = "Local reproduction scaffold for TTT-E2E"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "chex>=0.1.87,<0.2",
    "einops>=0.8.0,<0.9",
    "einx>=0.3.0,<0.4",
    "equinox>=0.11.12,<0.12",
    "grain>=0.2.7,<0.3",
    "jaxtyping>=0.2.36,<0.3",
    "jax[cuda12]>=0.4.34,<0.6; platform_system == 'Linux'",
    "jax>=0.4.34,<0.6; platform_system != 'Linux'",
    "matplotlib>=3.9,<3.11",
    "numcodecs>=0.15.1,<0.16",
    "numpy>=2.0,<2.2",
    "optax>=0.2.4,<0.3",
    "orbax-checkpoint>=0.11.13,<0.12",
    "tensorstore>=0.1.71,<0.2",
    "tqdm>=4.67.1,<5",
    "wandb>=0.19,<0.21",
    "zarr>=3.0.4,<3.2",
]

[project.scripts]
train = "ttt.train:main"

[tool.setuptools.packages.find]
include = ["ttt", "ttt.*", "hydra_plugins", "hydra_plugins.*"]
